fix: list, types
