+++ src/assets/restaurants.ts
@@ -1,23 +1,4 @@
-type Menu = {
-  id: string;
-  name: string;
-  price: number;
-  ingredients: string[];
-};
-
-type Review = {
-  id: string;
-  user: string;
-  text: string;
-  rating: number;
-};
-
-type Restaurant = {
-  id: string;
-  name: string;
-  menu: Menu[];
-  reviews: Review[];
-};
+import type { Restaurant } from "../types";
 
 export const restaurants: Restaurant[] = [
   {
