+++ src/types/index.ts
@@ -1,0 +1,20 @@
+type Menu = {
+  id: string;
+  name: string;
+  price: number;
+  ingredients: string[];
+};
+
+type Review = {
+  id: string;
+  user: string;
+  text: string;
+  rating: number;
+};
+
+export type Restaurant = {
+  id: string;
+  name: string;
+  menu: Menu[];
+  reviews: Review[];
+};
